+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 
 ## Development
 
+* 2025-04-15 -- 0.1.0 DLSite Scraper
+  * DLSite Scraper
+  * Views
+  * Database
+
 * 2025-03-11 -- 0.0.1 Project start
   * Project start